Removing all additional functionality
