--- css/print.css
@@ -1,10 +1,0 @@
-@media print {
-    .remove-when-printing {
-      display: none;
-    }
-}
-@media screen {
-    .remove-when-printing {
-      display: none;
-    }
-}
