--- firebase.json
@@ -1,5 +1,0 @@
-{
-  "database": {
-    "rules": "database.rules.json"
-  }
-}
