+++ js/compact/document.js
@@ -1,0 +1,12 @@
+let Document = {
+    text: [],
+    textLog: [],
+    body: [],
+    words: [],
+    wordKeys: [],
+    createTimeStamp: function(){
+        //creates a timeStamp
+        let date = new Date();
+        return date.getTime();
+    },
+}
