+++ js/compact/processor/words.js
@@ -1,0 +1,88 @@
+let Words = {
+    globalWords: [],
+    words: [],
+    wordCount: 0,
+    wordLengths: [],
+    syllables:[],
+    breaths: [],
+    reset: function(){
+        // Set all global data to empty
+        this.globalWords = [];
+        this.words = [];
+        this.wordCount = 0;
+        this.wordLengths = [];
+        this.syllables = [];
+        this.breaths = [];
+    },
+    checkWordValidity: function (word){
+        let areNumbers = word.match(/\d+/g);
+        if(areNumbers != null){
+            return false;
+        }else{
+            return true;
+        }
+    },
+    processWords: function(sentence, caller){
+        //remove all non-word characters
+        sentence = sentence.replace(/[\W_]+/g," ");
+        //Split words from sentence by spaces
+        let words = sentence.split(" ");
+        //Define a new Word Bank
+        let Bank = new Object;
+        //Initailize empty arrays for word lengths, how many, and the words themselves
+        Bank.lengths = [];
+        Bank.count = 0;
+        Bank.words = [];
+        Bank.syllables = [];
+        Bank.breaths = [];
+        //Initialize a wordCount variable
+        let wordCount = 0;
+        //For each of the words generated by the split
+        words.forEach(function(word){
+                //Continue if normal
+                if(Words.checkWordValidity(word) == true){
+                    if(word.length > 0){
+
+                        wordCount = wordCount + 1;
+                        let syllableCount = getSyllableCount(word);
+                        Bank.syllables.push(syllableCount);
+                        //Add a copy of the word in the word Bank
+                        Bank.words.push(word);
+                        //Add word lengths to the word Bank
+                        Bank.lengths.push(word.length);
+                        //Create breath unit variable
+                        let breath = (word.length/syllableCount).toFixed(2);
+                        //Add breath units to word Bank
+                        Bank.breaths.push(breath);
+                        //Add syllable counts to global Words
+                        Words.syllables.push(syllableCount);
+                        //Add word to the global Words
+                        Words.globalWords.push(word);
+                        //Add length to the global Words
+                        Words.wordLengths.push(word.length);
+                        //Add breath unit to global Words
+                        Words.breaths.push(breath);
+                        //Push word into Document word bank
+                        if(caller == "sentence"){
+                            //Push words to Document word bank
+                            Document.words.push(word);
+                            //Push reference key to word
+                            Document.wordKeys.push(createKey());
+                        }
+                    }
+                //Do not add word to word bank, putword in special case bank
+                }else{
+                    if(caller == "sentence"){
+                        // There's still a word here, so it counts, but can't be processed
+                        wordCount = wordCount + 1;
+                    }
+                }
+        });
+        //Add the number of words to the Bank
+        Bank.count = wordCount;
+        //Push copy of word bank into global word bank
+        Words.wordCount = wordCount;
+        //Return Bank to Sentences
+        return Bank;
+    },
+}
