--- js/dictionary.js
@@ -1,491 +1,0 @@
-let Corpora = {
-    honora: [
-      "Admiral",
-      "Adm",
-      "Air Cdre",
-      "Air Marshal",
-      "Alderman",
-      "Ambassador",
-      "Archbishop",
-      "Archdeacon",
-      "Baron of Lewes",
-      "Baron",
-      "Baroness",
-      "Bishop",
-      "Brigadier",
-      "Brother",
-      "Buergermeister",
-      "Canon",
-      "Captain",
-      "Catholic Chaplain",
-      "Chairman",
-      "Chevalier",
-      "Chief Constable",
-      "Chief Fire Officer",
-      "Chief Inspector",
-      "Chief Rabbi Sir",
-      "Chief Supt",
-      "Chief",
-      "Colonel Dr",
-      "Colonel",
-      "Commander",
-      "Commodore",
-      "Comte",
-      "Consort to the Deputy Mayor of Wirral",
-      "Consul",
-      "Councillor Mrs",
-      "Councillor",
-      "Count",
-      "Countess",
-      "Dame",
-      "Dato'",
-      "Datuk",
-      "Dean",
-      "Deputy Mayor of Sefton Councillor",
-      "Deputy Vice-Chancellor Professor",
-      "District Attorney",
-      "District Judge",
-      "Doctor",
-      "Dott.",
-      "Dr",
-      "Duchess",
-      "Duke",
-      "Earl",
-      "Emeritus Professor",
-      "Eng",
-      "Eur Ing",
-      "Esp",
-      "Father",
-      "Fiona",
-      "Flt Lt",
-      "Frau",
-      "Frú",
-      "General",
-      "Governor",
-      "Group Captain",
-      "Her Excellency",
-      "Her Honour Judge",
-      "Her Honour",
-      "Her Royal Highness Princess",
-      "Her Royal Highness",
-      "Herr",
-      "High Sheriff of Merseyside",
-      "His Excellency Chief",
-      "His Excellency",
-      "His Grace the Duke of Westminster",
-      "His Grace The Duke",
-      "His Grace",
-      "His Honour Judge",
-      "His Honour",
-      "His Royal Highness Prince",
-      "His Royal Highness",
-      "Hon. Lord",
-      "Hon",
-      "Honourable",
-      "HRH Princess",
-      "Ing",
-      "Jr",
-      "Junior",
-      "Judge",
-      "Justice",
-      "Lady Mayoress",
-      "Lady",
-      "Lieutenant Colonel",
-      "Lieutenant Commander",
-      "Lieutenant",
-      "Lord Lieutenant of Merseyside Dame",
-      "Lord Lieutenant of Merseyside Dr",
-      "Lord Lieutenant",
-      "Lord Mayor",
-      "Lord",
-      "Lt Cdr",
-      "Lt Colonel",
-      "Madam",
-      "Mhms",
-      "Mademoiselle",
-      "Major",
-      "Majors",
-      "Mar",
-      "Marchioness",
-      "Master",
-      "Mayor",
-      "Mayoress",
-      "Messrs",
-      "Methodist Chaplain",
-      "Miss",
-      "Mme",
-      "Monsieur",
-      "Monsignor",
-      "Mons",
-      "Mr and Mrs",
-      "Mr Justice",
-      "Mr",
-      "Mrs",
-      "Ms",
-      "Mx",
-      "of Birkenhead",
-      "of Quarry Bank",
-      "Pastor",
-      "Police Constable",
-      "Pr Eng",
-      "Prebendary",
-      "President",
-      "Prince",
-      "Princess",
-      "Pro-Vice-Chancellor Professor",
-      "Pro-Vice-Chancellor",
-      "Professor Dame",
-      "Professor Lord",
-      "Professor Sir",
-      "Professor the Lord",
-      "Professor",
-      "Prof",
-      "QPM",
-      "Rabbi",
-      "Revd. Cannon Dr.",
-      "Reverend Brother",
-      "Reverend Canon",
-      "Reverend Dr",
-      "Reverend Father",
-      "Reverend Professor",
-      "Reverend Sister",
-      "Reverend",
-      "Rev",
-      "Right Reverend",
-      "Rt Hon",
-      "Rt. Hon. Dame",
-      "Rt. Hon. Lord Justice",
-      "Rt. Hon. Sir",
-      "Senator",
-      "Senior Executive Assistant",
-      "Sergent",
-      "Sheikh",
-      "Sir",
-      "Sister",
-      "Squadron Leader",
-      "Sub Lt",
-      "Saint",
-      "St",
-      "Senior",
-      "Sr",
-      "The Countess of",
-      "The Duke of",
-      "The Earl of",
-      "The Editor",
-      "The High Sheriff of Merseyside Dr.",
-      "The High Sheriff",
-      "The Hon Dame",
-      "The Hon. Dr.",
-      "The Hon. Mr. Justice",
-      "The Hon. Mr",
-      "The Honorable Dame",
-      "The Honorable Mrs",
-      "The Honorable Sir",
-      "The Honorable",
-      "The Lord",
-      "The Mayor-Elec of Wirral Councillor",
-      "The Mayor",
-      "The Most Rev",
-      "The Right Hon. Lord Justice",
-      "The Right Hon. the Lord",
-      "The Right Honourable",
-      "The Rt Hon Lady Justice",
-      "The Rt Hon Lord",
-      "The Rt Hon Mrs",
-      "The Rt Hon the Lady",
-      "The Rt Hon the Lord",
-      "The Rt Hon",
-      "The Rt Honourable Earl of",
-      "The Rt Rev Bishop",
-      "The Rt Rev Dr",
-      "The Rt Rev",
-      "The Rt. Hon the Lord",
-      "The Rt. Hon. Dame",
-      "The Rt. Hon. Lady Justice",
-      "The Rt. Hon. Lord",
-      "The Rt. Hon. The Baroness",
-      "The Rt. Hon. the Earl of",
-      "The Rt. Hon. The Lady",
-      "The Rt. Hon. The Lord",
-      "The Rt. Hon.",
-      "The Rt. Hon",
-      "The Rt. Rev. Dr.",
-      "The Rt. Revd. Dr.",
-      "The Rt. Worshipful",
-      "The Rt. Worshipfull",
-      "The Very Revd. Dr.",
-      "The Very Reverend",
-      "The Worship the Mayor of Wirral",
-      "The Worship the Mayor",
-      "The Worshipful the Lord Mayor of Liverpool",
-      "The Worshipful the Lord Mayor",
-      "The Worshipful the Mayor of Halton",
-      "The Worshipful the Mayor of Knowsley",
-      "The Worshipful the Mayor of Sefton",
-      "The Worshipful the Mayor of St. Helens",
-      "The Worshipful the Mayor of",
-      "The Worshipful the Mayor",
-      "Venerable",
-      "Very Reverend",
-      "Vice Admiral Sir",
-      "Vice-Chancellor Professor",
-      "Vice-Chancellor",
-      "Viscount",
-      "Viscountess",
-      "Wing Commander",
-      "Mon",
-      "Tues",
-      "Wed",
-      "Thurs",
-      "Fri",
-      "Sat",
-      "Sun",
-      "Inc",
-      "Incorporated",
-    ],
-    replacement: [
-      "Admiral",
-      "Admiral",
-      "Air Comander",
-      "Air Marshal",
-      "Alderman",
-      "Ambassador",
-      "Archbishop",
-      "Archdeacon",
-      "Baron of Lewes",
-      "Baron",
-      "Baroness",
-      "Bishop",
-      "Brigadier",
-      "Brother",
-      "Buergermeister",
-      "Canon",
-      "Captain",
-      "Catholic Chaplain",
-      "Chairman",
-      "Chevalier",
-      "Chief Constable",
-      "Chief Fire Officer",
-      "Chief Inspector",
-      "Chief Rabbi Sir",
-      "Chief Supt",
-      "Chief",
-      "Colonel Dr",
-      "Colonel",
-      "Commander",
-      "Commodore",
-      "Comte",
-      "Consort to the Deputy Mayor of Wirral",
-      "Consul",
-      "Councillor Mrs",
-      "Councillor",
-      "Count",
-      "Countess",
-      "Dame",
-      "Dato'",
-      "Datuk",
-      "Dean",
-      "Deputy Mayor of Sefton Councillor",
-      "Deputy Vice-Chancellor Professor",
-      "District Attorney",
-      "District Judge",
-      "Doctor",
-      "Dottore.",
-      "Doctor",
-      "Duchess",
-      "Duke",
-      "Earl",
-      "Emeritus Professor",
-      "Engineer",
-      "European Engineer",
-      "Espquire",
-      "Father",
-      "Fiona",
-      "Flight Lieutenant",
-      "Frau",
-      "Frú",
-      "General",
-      "Governor",
-      "Group Captain",
-      "Her Excellency",
-      "Her Honour Judge",
-      "Her Honour",
-      "Her Royal Highness Princess",
-      "Her Royal Highness",
-      "Herr",
-      "High Sheriff of Merseyside",
-      "His Excellency Chief",
-      "His Excellency",
-      "His Grace the Duke of Westminster",
-      "His Grace The Duke",
-      "His Grace",
-      "His Honour Judge",
-      "His Honour",
-      "His Royal Highness Prince",
-      "His Royal Highness",
-      "Hononorable Lord",
-      "Honour",
-      "Honourable",
-      "Her Royal Highness Princess",
-      "Ing",
-      "Junior",
-      "Junior",
-      "Judge",
-      "Justice",
-      "Lady Mayoress",
-      "Lady",
-      "Lieutenant Colonel",
-      "Lieutenant Commander",
-      "Lieutenant",
-      "Lord Lieutenant of Merseyside Dame",
-      "Lord Lieutenant of Merseyside Doctor",
-      "Lord Lieutenant",
-      "Lord Mayor",
-      "Lord",
-      "Lieutenant Commander",
-      "Lieutenant Colonel",
-      "Madam",
-      "Madams",
-      "Mademoiselle",
-      "Major",
-      "Majors",
-      "Marchioness",
-      "Marchioness",
-      "Master",
-      "Mayor",
-      "Mayoress",
-      "Monsieurs",
-      "Methodist Chaplain",
-      "Miss",
-      "Madame",
-      "Monsieur",
-      "Monsignor",
-      "Monsieur",
-      "Mister and Miss",
-      "Mister Justice",
-      "Mister",
-      "Miss",
-      "Miss",
-      "Mx",
-      "of Birkenhead",
-      "of Quarry Bank",
-      "Pastor",
-      "Police Constable",
-      "Proffesional Engineer",
-      "Prebendary",
-      "President",
-      "Prince",
-      "Princess",
-      "Pro-Vice-Chancellor Professor",
-      "Pro-Vice-Chancellor",
-      "Professor Dame",
-      "Professor Lord",
-      "Professor Sir",
-      "Professor the Lord",
-      "Professor",
-      "Professor",
-      "Quenn's Police Metal",
-      "Rabbi",
-      "Revered Cannon Doctor",
-      "Reverend Brother",
-      "Reverend Canon",
-      "Reverend Doctor",
-      "Reverend Father",
-      "Reverend Professor",
-      "Reverend Sister",
-      "Reverend",
-      "Reverend",
-      "Right Reverend",
-      "Right Honorable",
-      "Right Honorable Dame",
-      "Right Honorable Lord Justice",
-      "Right Honorable Sir",
-      "Senator",
-      "Senior Executive Assistant",
-      "Sergent",
-      "Sheikh",
-      "Sir",
-      "Sister",
-      "Squadron Leader",
-      "Sub Lt",
-      "Saint",
-      "Saint",
-      "Senior",
-      "Senior",
-      "The Countess of",
-      "The Duke of",
-      "The Earl of",
-      "The Editor",
-      "The High Sheriff of Merseyside Doctor",
-      "The High Sheriff",
-      "The Honorable Dame",
-      "The Honorable Doctor",
-      "The Honorable Mister Justice",
-      "The Honorable Mister",
-      "The Honorable Dame",
-      "The Honorable Miss",
-      "The Honorable Sir",
-      "The Honorable",
-      "The Lord",
-      "The Mayor Electorate of Wirral Councillor",
-      "The Mayor",
-      "The Most Reverend",
-      "The Right Honourable Lord Justice",
-      "The Right Honourable the Lord",
-      "The Right Honourable",
-      "The Right Honourable Lady Justice",
-      "The Right Honourable Lord",
-      "The Right Honourable Mrs",
-      "The Right Honourable the Lady",
-      "The Right Honourable the Lord",
-      "The Right Honourable",
-      "The Right Honourable Earl of",
-      "The Right Rev Bishop",
-      "The Right Reverend Doctor",
-      "The Right Reverend",
-      "The Right Honourable the Lord",
-      "The Right Honourable Dame",
-      "The Right Honourable Lady Justice",
-      "The Right Honourable Lord",
-      "The Right Honourable The Baroness",
-      "The Right Honourable the Earl of",
-      "The Right Honourable The Lady",
-      "The Right Honourable The Lord",
-      "The Right Honourable",
-      "The Right Honourable",
-      "The Right Reverend Doctor",
-      "The Right Revered Doctor",
-      "The Right Worshipful",
-      "The Right Worshipfull",
-      "The Very Revered Doctor",
-      "The Very Reverend",
-      "The Worship the Mayor of Wirral",
-      "The Worship the Mayor",
-      "The Worshipful the Lord Mayor of Liverpool",
-      "The Worshipful the Lord Mayor",
-      "The Worshipful the Mayor of Halton",
-      "The Worshipful the Mayor of Knowsley",
-      "The Worshipful the Mayor of Sefton",
-      "The Worshipful the Mayor of Saint Helens",
-      "The Worshipful the Mayor of",
-      "The Worshipful the Mayor",
-      "Venerable",
-      "Very Reverend",
-      "Vice Admiral Sir",
-      "Vice-Chancellor Professor",
-      "Vice-Chancellor",
-      "Viscount",
-      "Viscountess",
-      "Wing Commander",
-      "Monday",
-      "Tuesday",
-      "Wednesday",
-      "Thursday",
-      "Friday",
-      "Saturday",
-      "Sunday",
-      "Incorporated",
-      "Incorporated",
-    ],
-
-}
