--- js/mobile/detectViewport.js
@@ -1,5 +1,0 @@
-(function(){
-    if(window.innerWidth < 500){
-        window.location.assign("mobile.html");
-    }
-})();
