--- js/mobile/mobileProcess.js
@@ -1,3 +1,0 @@
-function beginMobileProcess(){
-    window.location.assign("mobile-results.html");
-}
