--- js/mobile/navigation.js
@@ -1,38 +1,0 @@
-/**
-    Navigation from w3schools template
-**/
-
-// Get the Sidebar
-var mySidebar = document.getElementById("mySidebar");
-
-// Get the DIV with overlay effect
-var overlayBg = document.getElementById("myOverlay");
-
-// Toggle between showing and hiding the sidebar, and add overlay effect
-function w3_open() {
-    if (mySidebar.style.display === 'block') {
-        mySidebar.style.display = 'none';
-        overlayBg.style.display = "none";
-    } else {
-        mySidebar.style.display = 'block';
-        overlayBg.style.display = "block";
-    }
-}
-
-// Close the sidebar with the close button
-function w3_close() {
-    mySidebar.style.display = "none";
-    overlayBg.style.display = "none";
-}
-
-function relocateTest(){
-    window.location.assign("test.html");
-}
-
-function relocateUse(){
-    window.location.assign("use.html");
-}
-
-function relocateResults(){
-    window.location.assign("results.html");
-}
