--- js/utilities/log.js
@@ -1,10 +1,0 @@
-/**
-  Source:
-  https://codepen.io/ssk7833/pen/JKPYoX
-**/
-
-function log() {
-  setTimeout(console.log.bind(console, "%c Hello, I see you're interested in my code. If you would like to contact the developer feel free to contact me at: kenyan@kenyanburnham.com %c", "background: #007bff;color:#FFF;padding:5px;border-radius: 5px;line-height: 26px;", ""));
-}
-
-log();
